First Commit - Front Page
This has the basic framework of the front page with a dummy header and footer. The images on the front page are in public, and the code is in App.js. The images in the carousel cannot be clicked on - needs work. The images not in the carousel can be clicked on, but they currently do not go anywhere (will go to future pages). Continuing to work on the images by adding text to describe where clicking on the image will take you.